# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Clients & Comptes",
        "description": "API RESTful pour la gestion des clients et de leurs comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister tous les utilisateurs",
                "description": "Récupère une liste paginée d'utilisateurs avec possibilité de filtrage et tri",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filtrer par rôle",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "client"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par nom, email ou téléphone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "dateCreation",
                            "enum": [
                                "dateCreation",
                                "nom"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                },
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur",
                "description": "Crée un nouvel utilisateur dans le système",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "nci",
                                    "email",
                                    "telephone",
                                    "adresse"
                                ],
                                "properties": {
                                    "nom": {
                                        "description": "Nom complet",
                                        "type": "string"
                                    },
                                    "nci": {
                                        "description": "Numéro CNI",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Adresse email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telephone": {
                                        "description": "Numéro de téléphone",
                                        "type": "string"
                                    },
                                    "adresse": {
                                        "description": "Adresse complète",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Rôle utilisateur",
                                        "type": "string",
                                        "default": "client",
                                        "enum": [
                                            "admin",
                                            "client"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur créé avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "properties": {
                    "currentPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    },
                    "itemsPerPage": {
                        "type": "integer"
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Amadou Diallo"
                    },
                    "nci": {
                        "type": "string",
                        "example": "123456789012"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "amadou.diallo@email.com"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221771234567"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "Dakar, Sénégal"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "client"
                        ],
                        "example": "client"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "derniereModification": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}